feat : MainHUD
- 메인 화면에서 사용하는 MainHUD 생성
- InGameMode 변수 타입 수정

+++ Source/PullAndPush/Private/Game/InGameInstance.cpp
@@ -13,7 +13,7 @@ bool UInGameInstance::IsFirstRoundStart()
 {
 	return bIsFirstRoundStart;
 }
-void UInGameInstance::InitSetting(const int InMaxRoundCount)
+void UInGameInstance::InitSetting(const int8 InMaxRoundCount)
 {
 	MaxRoundCount = InMaxRoundCount;
 	bIsFirstRoundStart = true;
